refactor(core): replace regions with marks
A mark represents a single position in the buffer,
whereas a region was a range that was considered a cursor when
it had length 0.
In most cases, at least so far, we've been dealing with cursors,
so this simplifies the code a good bit, and gets rid of a few
hard-to-encode-in-types invariants.

A future plan would be to represent any regions / selections through
tuples of marks.